Otel_API/requirements.txt
opentelemetry-proto == 1.34.1
